Add 6-month and 1-year lookback presets
Data is monthly closes, so 1d/1w/1M/3M can't produce a meaningful
backtest (too few points); added the two short windows that can — 6
months (7 points) and 1 year (13). lookbackYears() now uses parseFloat
so the fractional 0.5y (6-month) preset works; everything downstream
(buildMonthAxis, SPUS gating) already handles fractional years.

Verified in-browser: 6mo -> 7-point window, 1y -> 13, default still 5y;
SPUS stays enabled for both; no console errors; tests 17/17.

+++ src/app.js
@@ -80,10 +80,11 @@ function loadSeries(tkr){
   if(b && b.length){ priceCache[tkr] = b; return b; }
   return null;
 }
-// Effective lookback in years: a preset, or the "Custom…" input. 0 = "Max available".
+// Effective lookback in years: a preset (may be fractional, e.g. 0.5 = 6 months),
+// or the "Custom…" input. 0 = "Max available".
 function lookbackYears(){
   if($("#years").value === "custom"){ const n = parseInt($("#yearsCustom").value); return n > 0 ? n : 0; }
-  return parseInt($("#years").value);
+  return parseFloat($("#years").value);
 }
 // SPUS launched Dec 2019, so it can't cover a lookback longer than its own history.
 // When the selected lookback reaches back past SPUS's first month (or "Max available"),
